+++ picosentry/__init__.py
@@ -14,4 +14,4 @@ Usage:
     picosentry serve
 """
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"